fixati bug finali, test completo dell'applicazione fatto, tutte le funzioni sistemate

+++ app/src/main/java/com/example/gestorednd/DataClasses/Spells.kt
@@ -1,4 +1,4 @@
 package com.example.gestorednd.DataClasses
 
-data class Spells (var name : String, var lvl : Int) {
+data class Spells(var name: String, var lvl: Long) {
 }
